Removed ATM, Teller classes
GUI Functioning
